+++ CHANGELOG.md
@@ -7,10 +7,14 @@
 ### [22/3/22] SPLOG Basics
 * Constructed Splog framework  
 * Can now log strings properly  
-* **TODO** Enable variable logging  
-* **TODO** Enable setting up logging format  
-* **TODO** Enable variant color for different levels of logging
+* [x] **TODO** Enable variable logging  
+* [x] **TODO** Enable setting up logging format  
+* [x] **TODO** Enable variant color for different levels of logging
 
 ### [22/3/23] SPLOG Improvements
 * **FIX** Variable logging
 * **FIX** Colorize message (only for whole line)
+
+### [22/3/25] SPLOG Improvements
+* **FIX** Colorizing segment of line
+* **FIX** Setting up logging format